rename `features` -> `featurizers`
