+++ chemprop/args.py
@@ -12,7 +12,7 @@ from tap import Tap  # pip install typed-argument-parser (https://github.com/swa
 
 import chemprop.data.utils
 from chemprop.data import set_cache_mol, empty_cache
-from chemprop.features import get_available_features_generators
+from chemprop.featurizers import get_available_features_generators
 
 
 Metric = Literal['auc', 'prc-auc', 'rmse', 'mae', 'mse', 'r2', 'accuracy', 'cross_entropy', 'binary_cross_entropy', 'sid', 'wasserstein', 'f1', 'mcc', 'bounded_rmse', 'bounded_mae', 'bounded_mse']
