+++ chemprop/data/utils.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 from .data import MoleculeDatapoint, MoleculeDataset, make_mols
 from .scaffold import log_scaffold_stats, scaffold_split
 from chemprop.args import PredictArgs, TrainArgs
-from chemprop.features import load_features, load_valid_atom_or_bond_features, is_mol
+from chemprop.featurizers import load_features, load_valid_atom_or_bond_features, is_mol
 
 def get_header(path: str) -> List[str]:
     """
