+++ chemprop/models/mpn.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn as nn
 
 from chemprop.args import TrainArgs
-from chemprop.features import BatchMolGraph, get_atom_fdim, get_bond_fdim, mol2graph
+from chemprop.featurizers import BatchMolGraph, get_atom_fdim, get_bond_fdim, mol2graph
 from chemprop.nn_utils import index_select_ND, get_activation_function
