+++ chemprop/sklearn_predict.py
@@ -6,7 +6,7 @@ from tqdm import tqdm
 
 from chemprop.args import SklearnPredictArgs, SklearnTrainArgs
 from chemprop.data import get_data
-from chemprop.features import get_features_generator
+from chemprop.featurizers import get_features_generator
 from chemprop.sklearn_train import predict
 from chemprop.utils import makedirs, timeit
