+++ scripts/save_features.py
@@ -12,7 +12,7 @@ from tap import Tap  # pip install typed-argument-parser (https://github.com/swa
 sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
 
 from chemprop.data import get_smiles
-from chemprop.features import get_available_features_generators, get_features_generator, load_features, save_features
+from chemprop.featurizers import get_available_features_generators, get_features_generator, load_features, save_features
 from chemprop.utils import makedirs
